New traces

+++ src/main/java/hpdp/storage/Storage.java
@@ -5,6 +5,8 @@ import java.io.IOException;
 import java.util.List;
 import java.util.ArrayList;
 import java.io.BufferedWriter;
+import java.io.BufferedReader;
+import java.io.FileReader;
 import java.io.FileWriter;
 import java.io.File;
 
@@ -12,7 +14,9 @@ import java.io.File;
 public class Storage{
 
     BufferedWriter file;
+    BufferedReader fileReader;
     public String filePath = null;
+    public boolean read = false;
 
     public Storage(){
         filePath = null;
@@ -22,9 +26,16 @@ public class Storage{
 
     public Storage newFile(String filePath) throws IOException{
 
-        File existingFile = new File(filePath);
-        existingFile.delete();
-        file = new BufferedWriter(new FileWriter(filePath));
+
+
+        if (read) {
+            fileReader = new BufferedReader(new FileReader(filePath));
+        }
+        else{
+            File existingFile = new File(filePath);
+            existingFile.delete();
+            file = new BufferedWriter(new FileWriter(filePath));
+        }
 
 
         return this;
@@ -37,6 +48,17 @@ public class Storage{
             file.flush();
         }
 
+    }
+
+    public String readLine() throws IOException {
+        if (fileReader != null){
+            return fileReader.readLine();
+        }
+        return null;
+    }
+
+    public void setRead(boolean read){
+        this.read = read;
     }
